pipelines/ingest/requirements.txt
# Seedy Autoingesta Pipeline
pydantic-settings>=2.0
pyyaml>=6.0
httpx>=0.27
feedparser>=6.0
tenacity>=8.2
trafilatura>=1.12
pymupdf>=1.24
qdrant-client>=1.12
